add and remove buttons working
