+++ app/src/app.js
@@ -9,3 +9,7 @@ const app = angular.module('imageGallery', [
 const dev = 'http://localhost:3000/api';
 
 app.value('apiUrl', dev);
+
+app.factory('apiUrl', function() {
+  return dev;
+});
