+++ app/src/components/add_img/add-img.html
@@ -1,0 +1,15 @@
+<section>
+  <div>
+    <label>Title:</label>
+    <input ng-model="$ctrl.name">
+  </div>
+  <div>
+    <label>Image Source (URL):</label>
+    <input ng-model="$ctrl.src">
+  </div>
+  <div>
+    <label>Description:</label>
+    <input ng-model="$ctrl.description">
+  </div>
+  <button ng-click="$ctrl.addNew()">Add Image</button>
+</section>
