+++ app/src/components/add_img/add-img.js
@@ -1,0 +1,19 @@
+import template from './add-img.html';
+
+export default {
+  template,
+  bindings: {
+    add: '<'
+  },
+  controller
+};
+
+function controller() {
+  this.addNew = () => {
+    this.add({
+      name: this.name,
+      src: this.src,
+      description: this.description
+    });
+  };
+}
