+++ app/src/components/gallery_views/gallery-views.html
@@ -1,4 +1,5 @@
 <section>
+  <add-img add="$ctrl.add"></add-img>
   <div>
     <p>How would you like your images to be displayed?</p>
     <select ng-model="viewMenu" ng-options="view for view in $ctrl.views">
