+++ app/src/components/thumbnail/thumbnail-app.html
@@ -1,8 +1,6 @@
 <section>
-  <h3>Thumbnail View</h3>
-    <div ng-repeat="image in $ctrl.images">
-    </div>
-  <div>
-    <img ng-src="{{$ctrl.thumbnail.src}}" ng-class="$ctrl.styles.img">
+  <div ng-repeat="images in $ctrl.images">
+    <img ng-src="{{images.src}}" ng-class="$ctrl.styles.img">
+    <button ng-click="$ctrl.remove(images)">Delete</button>
   </div>
 </section>
