+++ app/src/components/title_link_desc/title-link-desc.html
@@ -1,5 +1,6 @@
- <div>
-    <h1>{{$ctrl.title}}</h1>
-    <p><a href="{{$ctrl.link.src}}">{{$ctrl.link.src}}</a></p>
-    <p>{{$ctrl.desc}}</p>
+ <div ng-repeat="images in $ctrl.images">
+    <h1>{{images.name}}</h1>
+    <p><a href="{{images.src}}">{{images.src}}</a></p>
+    <p>{{images.description}}</p>
+    <button ng-click="$ctrl.remove(images)">Delete</button>
 </div>
