+++ app/src/components/title_link_desc/title-link-desc.js
@@ -2,16 +2,22 @@ import template from './title-link-desc.html';
 
 export default {
   template,
-  bindings: {
-    bunny: '='
-  },
   controller
 };
 
-function controller() {
-  this.link = {
-    src: 'http://f.cl.ly/items/3g3J1G0w122M360w380O/3726490195_f7cc75d377_o.jpg'
+controller.$inject = ['imageService'];
+
+function controller(images) {
+  images.get().then(images => {
+    console.log(images);
+    this.images = images;
+  });
+
+  this.remove = image => {
+    images.remove(image._id)
+      .then(() => {
+        const index = this.images.indexOf(image);
+        if(index > -1) this.images.splice(index, 1);
+      });
   };
-  this.title = 'Cute Bunny Picture';
-  this.desc = 'Picture of a cute bunny';
 }
